requirements.txt
pydantic>=2.6.0
python-dotenv>=1.0.0
groq>=0.11.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
# Web UI (optional) — only imported by app/server.py; CLI does not need these
fastapi>=0.110
uvicorn[standard]>=0.27
